// src/app/model/transaction-history.ts
export class TransactionHistory {
    transaction_id: string;
    //transaction_date: string;
    order_uid: string;
    category_id: string;
    total_amount: string;
    order_state: string;
    order_cbs: string;
    order_oid: string;
    order_utr: string;
    status_desc: string;
    category_name: string;
    //coupon_code: string | null;
    //subscriber_no: string | null;
    //company_name: string;
    user_uid: string;
    debit_acc_no: string;
    txn_mode: string;
    credit_acc_no: string;
    payment_initiated_at: string;
    payment_success_at: string;
    status: number;
//     updated_date: string;
//     notiid: string;
    //mbs_ref_number: string;
//     inserted_date: string;
}